modification done

+++ src/main/java/com/aye/issueTracker/service/DepartmentServiceImpl.java
@@ -1,14 +1,23 @@
 package com.aye.issueTracker.service;
 
 import com.aye.issueTracker.exception.InvalidDepartmentDataException;
+import com.aye.issueTracker.exception.NotDeletableException;
 import com.aye.issueTracker.exception.ResourceNotFoundException;
 import com.aye.issueTracker.model.Department;
+import com.aye.issueTracker.model.User;
 import com.aye.issueTracker.repository.DepartmentRepository;
+import com.aye.issueTracker.repository.EmployeeRepository;
+import com.aye.issueTracker.repository.TicketRepository;
+import com.aye.issueTracker.repository.UserRepository;
 import com.aye.issuetrackerdto.entityDto.DepartmentDto;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Lazy;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -18,11 +27,8 @@ public class DepartmentServiceImpl implements DepartmentService{
 
     @Autowired
     private DepartmentRepository departmentRepository;
-
     @Autowired
     private ModelMapper modelMapper;
-
-
 
     @Override
     public List<DepartmentDto> getAllDepartments() {
@@ -54,6 +60,11 @@ public class DepartmentServiceImpl implements DepartmentService{
 
         departmentDto.setId(sequenceGeneratorService.generateSequence(Department.SEQUENCE_NAME));
         Department department = convertToEntity(departmentDto);
+
+        String createdByUserName = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Long createdByUserId = userRepository.findByUsername(createdByUserName).get().getId();
+        department.setCreatedById(createdByUserId);
+        department.setCreatedDateTime(LocalDateTime.now());
         department = departmentRepository.save(department);
         return converToDto(department);
     }
@@ -61,7 +72,15 @@ public class DepartmentServiceImpl implements DepartmentService{
     @Override
     public void deleteDepartment(Long id) throws ResourceNotFoundException {
         Optional<Department> department = departmentRepository.findById(id);
+
         if(department.isPresent()){
+            Boolean employeeExist = employeeRepository.existsByDepartment(department.get());
+            Boolean ticketExist = ticketRepository.existsByDepartment(department.get());
+
+            if (employeeExist || ticketExist) {
+                throw new NotDeletableException("Delete operation not possible! Associate Entity Exist!");
+            }
+
             departmentRepository.deleteById(id);
         }
         else {
@@ -80,10 +99,11 @@ public class DepartmentServiceImpl implements DepartmentService{
 
             if(departmentDto.getName() != null){
 
-                Optional<List<Department>> createdDepartments = departmentRepository.findByName(departmentDto.getName());
-                List<Department> departments = createdDepartments.get();
-                if (!departments.isEmpty()){
-                    throw new InvalidDepartmentDataException("Department Name Already Exist!");
+                Optional<Department> createdDepartment = departmentRepository.findByName(departmentDto.getName());
+                if (createdDepartment.isPresent()) {
+                    if (createdDepartment.get().getId() != departmentDto.getId()) {
+                        throw new InvalidDepartmentDataException("Department Name Already Exist!");
+                    }
                 }
 
                 if(departmentDto.getName() == null || departmentDto.getName().isEmpty()){
@@ -92,6 +112,11 @@ public class DepartmentServiceImpl implements DepartmentService{
 
                 department.get().setName(departmentDto.getName());
                 department.get().setDescription(departmentDto.getDescription());
+                String currentUserName = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+                Long currentUserId = userRepository.findByUsername(currentUserName).get().getId();
+                department.get().setUpdatedById(currentUserId);
+                department.get().setUpdatedDateTime(LocalDateTime.now());
+
                 Department updatedDepartment = departmentRepository.save(department.get());
                 return modelMapper.map(updatedDepartment, DepartmentDto.class);
             }
@@ -117,6 +142,12 @@ public class DepartmentServiceImpl implements DepartmentService{
 
     @Autowired
     private SequenceGeneratorService sequenceGeneratorService;
+    @Autowired
+    private EmployeeRepository employeeRepository;
+    @Autowired
+    private TicketRepository ticketRepository;
+    @Autowired
+    private UserRepository userRepository;
 
     @Override
     public void validDepartmentDto(DepartmentDto departmentDto) throws InvalidDepartmentDataException
@@ -133,9 +164,8 @@ public class DepartmentServiceImpl implements DepartmentService{
             throw new InvalidDepartmentDataException("Department Description Can't Be empty");
         }
 
-        Optional<List<Department>> createdDepartments = departmentRepository.findByName(departmentDto.getName());
-        List<Department> departments = createdDepartments.get();
-        if (!departments.isEmpty()){
+        Optional<Department> createdDepartment = departmentRepository.findByName(departmentDto.getName());
+        if (createdDepartment.isPresent()){
             throw new InvalidDepartmentDataException("Department Name Already Exist!");
         }
